{"schema":"decodex/commit/1","summary":"Modularize execution failure writeback domains","authority":"XY-1115"}
